+++ Plugins/Player/Player.hpp
@@ -51,6 +51,7 @@ private:
     ArgumentStack GetLanguage                       (ArgumentStack&& args);
     ArgumentStack SetResManOverride                 (ArgumentStack&& args);
     ArgumentStack SetCustomToken                    (ArgumentStack&& args);
+    ArgumentStack SetCreatureNameOverride           (ArgumentStack&& args);
 
     CNWSPlayer *player(ArgumentStack& args);